PersonalStockMarcket/Python_BackEnd/requirements.txt
django~=4.0.2
pymongo~=3.12.1
djongo==1.3.6
djangorestframework~=3.13.1
requests~=2.27.1
django-extensions==3.1.5